refactor(ai-blog-writer): decompose itinerary stops panel
